My_Flask_app/requirements.txt
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
google-auth==2.0.2
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient==2.1.0
npm==0.1.1
oauthlib==3.1.1
optional-django==0.1.0
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
